Add changes from code review
